Get rid of circular dependencies
