+++ src/reanimated2/commonTypes.ts
@@ -145,6 +145,14 @@ export interface Animation<T extends AnimationObject> extends AnimationObject {
   ) => void;
 }
 
+export enum SensorType {
+  ACCELEROMETER = 1,
+  GYROSCOPE = 2,
+  GRAVITY = 3,
+  MAGNETIC_FIELD = 4,
+  ROTATION = 5,
+}
+
 export interface NumericAnimation {
   current?: number;
 }
